requirements.txt
asgiref==3.7.2
autopep8==2.0.4
crispy-bootstrap4==2023.1
Django==4.2
django-ckeditor==6.7.0
django-cors-headers==4.3.1
django-crispy-forms==2.1
django-crontab==0.7.1
django-dbbackup==4.0.2
django-filter==23.5
django-js-asset==2.2.0
django-silk==5.0.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
extras==1.0.0
gprof2dot==2022.7.29
install==1.3.5
mysqlclient==2.2.1
numpy==1.26.3
pandas==2.1.4
pillow==10.2.0
pycodestyle==2.11.1
PyJWT==2.8.0
python-dateutil==2.8.2
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4
